pyproject.toml
[tool.poetry]
name = "python_mobile_1"
version = "0.1.0"
description = "Mobile tests part 1"
authors = ["vasvap <vasvap@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Appium-Python-Client = "^2.6.1"
selene ={version = "^2.0.0-beta.10", allow-prereleases = true}
pytest = "^7.1.3"
allure-pytest = "^2.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
